pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "passlib>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
